feat: added basic timecard command and functionality
Added ConsoleCommand `timecard` which allows user to see total hours in
past working week, and percentages for each task. Does not currently
distribute incidental time as a weighted function of time per task.

Noted an error in the database initialization that overwrites the
'tasks' and 'categories' tables, and could corrupt previous data.