fix: enforce unique ids and keys a for tasks table
Add a `db_populate_user_table` function, and use it to populate tasks
table from database. Tasks table includes all task key-name pairs that
have historically been used, with a unique id for each pair. The tasks
table will always reflect what the user specifies in titr.cfg - if a new
task is added, it will be updated in the table. If a new task is added
with an existing key, the key for the old task will be set to `null` and
the new task should be added.

The above should avoid corruption of the database by rearranging task
primary keys.

This needs to also be done to the categories table, and may take more
refactoring, since 'id' and 'key' have the same function.