fix: database initialized with ConsoleSession
Added `db_connection` as a member of ConsoleSession. To make this work
effectively __enter__ and __exit__ methods for ConsoleSession were
added; the __exit__ method closes the database connection. This change
involved a lot of refactoring of existing database functions, so that
the db_connection is no longer passed between them.