feat: incidental tasks are user configurable
And also show up in the timecard. Configure incidental tasks in ~/.titr/titr.cfg by entering the keys of tasks that are incidental as a comma separated list.

Timecard now lists hours and adjusted hours columns - Incidental time is
always 0 in adjusted hours and percentage, and is applied as a weighted
average to non-incidental tasks.