fix: replace Value and TypeError with InputError
Replaced remainder of raise ValueError and raise TypeError statements
with raise InputError, when the exception arises from the user entering
bad data or an unrecognized command. This way datum_console will handle
these exceptions and keep running, whereas it will crash with a full
stack trace if an error is due to the programming implementation.

Updated unit tests to accommodate, all passing.