refactor: Make TimeEntry independent dataclass
TimeEntry has been made into an independent dataclass. This means that a
TimeEntry object can be created without a ConsoleSession as its parent.

A new add_entry method was added to ConsoleSession, which assigns
comment & task strings to the entry, as well as assigning default values
based on the console configuration.

Many tests were refactored and re-run along with this change.