refactor: add _fetch_first function
I kept running into the problem where execute an sql query and use
cursor.fetchone() to get the result. The latter returns a tuple, usually
in the form of (1,), and I only want the first element. However
cursor.fetchone()[0] causes failue if hte result is None. This function
shortcuts this, returning either None or the 0th index of the query
result.