refactor: move outlook function to separate module.
Moved outlook code into outlook.py. This removes the pywintypes and
win32com.client imports from `titr_main.py`. Simplified the check for
`OUTLOOK_ENABLED` by simply checking if we're in Windows or not. Working
on my windows distribution. Not yet checked in Linux to see if this
crashed anything or not.

Addresses but does not close
https://github.com/blairfrandeen/titr/issues/3